// src/config/env.d.ts
export declare const env: {
    PORT: string;
    NODE_ENV: "development" | "production" | "test";
    GEMINI_API_KEY: string;
    VOICE_PROVIDER: string;
    VOICE_API_KEY: string;
    WHATSAPP_PROVIDER: string;
    WHATSAPP_TOKEN: string;
    TWILIO_ACCOUNT_SID?: string | undefined;
    TWILIO_WHATSAPP_NUMBER?: string | undefined;
    REDIS_URL: string;
    DATABASE_URL: string;
    TARGET_PHONE_NUMBER: string;
    MY_PHONE_NUMBER: string;
    RESUME_URL: string;
    ARCHITECTURE_IMAGE_URL: string;
};
//# sourceMappingURL=env.d.ts.map